Refactor legacy classes location and add text centering feature
Relocated 'LegacyColorProcessor' and 'Legacy' classes from the 'util.legacy' package to 'adventure.legacy' to improve code organization as suggested in a code review. This required updating the import statements in multiple classes. Additionally, introduced a new feature to support center-aligned text. To implement this, added 'CenterTagResolver' class to 'core.adventure.resolver' package. To support this feature, 'pixel-width' libraries (core and utils) were added in the build file.
