+++ eternalcore-core/src/main/java/com/eternalcode/core/feature/essentials/container/DisposalCommand.java
@@ -6,7 +6,7 @@ import com.eternalcode.core.notice.NoticeService;
 import com.eternalcode.core.translation.Translation;
 import com.eternalcode.core.translation.TranslationManager;
 import com.eternalcode.core.user.UserManager;
-import com.eternalcode.core.util.legacy.Legacy;
+import com.eternalcode.core.adventure.legacy.Legacy;
 import dev.rollczi.litecommands.command.execute.Execute;
 import dev.rollczi.litecommands.command.permission.Permission;
 import dev.rollczi.litecommands.command.route.Route;
