+++ eternalcore-core/src/main/java/com/eternalcode/core/feature/essentials/container/InventoryOpenCommand.java
@@ -2,7 +2,7 @@ package com.eternalcode.core.feature.essentials.container;
 
 import com.eternalcode.annotations.scan.command.DescriptionDocs;
 import com.eternalcode.core.notice.NoticeService;
-import com.eternalcode.core.util.legacy.Legacy;
+import com.eternalcode.core.adventure.legacy.Legacy;
 import dev.rollczi.litecommands.argument.Arg;
 import dev.rollczi.litecommands.command.execute.Execute;
 import dev.rollczi.litecommands.command.permission.Permission;
