+++ eternalcore-core/src/main/java/com/eternalcode/core/feature/essentials/item/ItemLoreCommand.java
@@ -3,7 +3,7 @@ package com.eternalcode.core.feature.essentials.item;
 import com.eternalcode.annotations.scan.command.DescriptionDocs;
 import com.eternalcode.core.notice.NoticeService;
 import com.eternalcode.core.util.AdventureUtil;
-import com.eternalcode.core.util.legacy.Legacy;
+import com.eternalcode.core.adventure.legacy.Legacy;
 import dev.rollczi.litecommands.argument.Arg;
 import dev.rollczi.litecommands.argument.joiner.Joiner;
 import dev.rollczi.litecommands.command.amount.Min;
